Get rid of Builder, which is redundant.
