+++ tests/JsonTokenTest.php
@@ -2,7 +2,7 @@
 declare(strict_types=1);
 namespace ParagonIE\PAST\Tests;
 
-use ParagonIE\PAST\Builder;
+use ParagonIE\PAST\JsonToken;
 use ParagonIE\PAST\Exception\PastException;
 use ParagonIE\PAST\Keys\{
     AsymmetricSecretKey,
@@ -11,10 +11,10 @@ use ParagonIE\PAST\Keys\{
 use PHPUnit\Framework\TestCase;
 
 /**
- * Class BuilderTest
+ * Class JsonTokenTest
  * @package ParagonIE\PAST\Tests
  */
-class BuilderTest extends TestCase
+class JsonTokenTest extends TestCase
 {
     /**
      * @covers Builder::getToken()
@@ -23,7 +23,7 @@ class BuilderTest extends TestCase
     public function testAuthDeterminism()
     {
         $key = new SymmetricAuthenticationKey('YELLOW SUBMARINE, BLACK WIZARDRY');
-        $builder = (new Builder())
+        $builder = (new JsonToken())
             ->setPurpose('auth')
             ->setKey($key)
             ->set('data', 'this is a signed message')
@@ -31,18 +31,18 @@ class BuilderTest extends TestCase
 
         $this->assertSame(
             'v2.auth.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAzOS0wMS0wMVQwMDowMDowMCswMDowMCJ9pBzDGiV6cmCX8Wxuj09xmNYiVOcD9yuE0TntviAEWvs=',
-            (string) $builder->getToken(),
+            (string) $builder,
             'Auth, no footer'
         );
         $footer = (string) \json_encode(['key-id' => 'gandalf0']);
         $this->assertSame(
             'v2.auth.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAzOS0wMS0wMVQwMDowMDowMCswMDowMCJ9X7CSuUFMklmXhHsu4bXvx_wiH_OyxpMijfNCjHiklXk=.eyJrZXktaWQiOiJnYW5kYWxmMCJ9',
-            (string) $builder->setFooter($footer)->getToken(),
+            (string) $builder->setFooter($footer),
             'Auth, footer'
         );
         $this->assertSame(
             'v2.auth.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAzOS0wMS0wMVQwMDowMDowMCswMDowMCJ9pBzDGiV6cmCX8Wxuj09xmNYiVOcD9yuE0TntviAEWvs=',
-            (string) $builder->setFooter('')->getToken(),
+            (string) $builder->setFooter(''),
             'Auth, removed footer'
         );
 
@@ -51,17 +51,17 @@ class BuilderTest extends TestCase
                 ->setKey(new AsymmetricSecretKey('YELLOW SUBMARINE, BLACK WIZARDRY'), true);
         $this->assertSame(
             'v2.sign.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAzOS0wMS0wMVQwMDowMDowMCswMDowMCJ9J_YYSKaRss60fvAT5jR0KzGXRrgAT9hSPbJ8O5w1F_EgTTU4Zbjms_ngcM_gEtUlvDIFs1QFv3GWkU4Ya6z9CQ==',
-            (string) $builder->getToken(),
+            (string) $builder,
             'Sign, no footer'
         );
         $this->assertSame(
             'v2.sign.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAzOS0wMS0wMVQwMDowMDowMCswMDowMCJ97mfQNGcQvr0ygbUO4XwAgTtb0NDlSLQpbG0hQRIr1JAo3XwUIq1HgSy84iA1h5fds-ZqwFkn1baN0WapHLwIAQ==.eyJrZXktaWQiOiJnYW5kYWxmMCJ9',
-            (string) $builder->setFooter($footer)->getToken(),
+            (string) $builder->setFooter($footer),
             'Sign, footer'
         );
         $this->assertSame(
             'v2.sign.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAzOS0wMS0wMVQwMDowMDowMCswMDowMCJ9J_YYSKaRss60fvAT5jR0KzGXRrgAT9hSPbJ8O5w1F_EgTTU4Zbjms_ngcM_gEtUlvDIFs1QFv3GWkU4Ya6z9CQ==',
-            (string) $builder->setFooter('')->getToken(),
+            (string) $builder->setFooter(''),
             'Sign, removed footer'
         );
     }
@@ -73,7 +73,7 @@ class BuilderTest extends TestCase
     {
         $key = new SymmetricAuthenticationKey('YELLOW SUBMARINE, BLACK WIZARDRY');
         $footerArray = ['key-id' => 'gandalf0'];
-        $builder = (new Builder())
+        $builder = (new JsonToken())
             ->setPurpose('auth')
             ->setKey($key)
             ->set('data', 'this is a signed message')
@@ -81,12 +81,12 @@ class BuilderTest extends TestCase
             ->setFooterArray($footerArray);
         $this->assertSame(
             'v2.auth.eyJkYXRhIjoidGhpcyBpcyBhIHNpZ25lZCBtZXNzYWdlIiwiZXhwIjoiMjAzOS0wMS0wMVQwMDowMDowMCswMDowMCJ9X7CSuUFMklmXhHsu4bXvx_wiH_OyxpMijfNCjHiklXk=.eyJrZXktaWQiOiJnYW5kYWxmMCJ9',
-            (string) $builder->getToken(),
+            (string) $builder,
             'Auth, footer'
         );
         $this->assertEquals(
             $footerArray,
-            $builder->getToken()->getFooterArray()
+            $builder->getFooterArray()
         );
     }
 }
